Fix bugs to resolve debug DDP training experiments
