+++ cosmos_transfer1/diffusion/config/base/net.py
@@ -40,10 +40,3 @@ FADITV2Config: LazyDict = L(GeneralDIT)(
     use_adaln_lora=True,
     adaln_lora_dim=256,
 )
-
-DebugDITConfig: LazyDict = copy.deepcopy(FADITV2Config)
-DebugDITConfig.patch_spatial = 2
-DebugDITConfig.patch_temporal = 1
-DebugDITConfig.model_channels = 128 * 2
-DebugDITConfig.num_blocks = 4
-DebugDITConfig.num_heads = 2
