+++ cosmos_transfer1/diffusion/config/training/tokenizer.py
@@ -19,8 +19,10 @@ from cosmos_transfer1.diffusion.training.modules.pretrained_vae import (
     JITVAE,
     JointImageVideoSharedJITTokenizer,
     VideoJITTokenizer,
+    DummyJointImageVideoTokenizer,
 )
 from cosmos_transfer1.utils.lazy_config import LazyCall as L
+from cosmos_transfer1.utils.lazy_config import LazyDict
 
 TOKENIZER_OPTIONS = {}
 
@@ -32,6 +34,8 @@ def tokenizer_register(key):
 
     return decorator
 
+
+DebugTokenizerConfig: LazyDict = L(DummyJointImageVideoTokenizer)(name="dummy_joint_image_video")
 
 @tokenizer_register("cosmos_diffusion_tokenizer_comp8x8x8")
 def get_cosmos_diffusion_tokenizer_comp8x8x8(resolution: str, chunk_duration: int) -> omegaconf.dictconfig.DictConfig:
