+++ cosmos_transfer1/diffusion/config/transfer/net_ctrl.py
@@ -15,7 +15,7 @@
 
 import copy
 
-from cosmos_transfer1.diffusion.config.base.net import FADITV2Config, DebugDITConfig
+from cosmos_transfer1.diffusion.config.base.net import FADITV2Config
 from cosmos_transfer1.diffusion.networks.general_dit_ctrl_enc import GeneralDITEncoder
 from cosmos_transfer1.diffusion.networks.general_dit_ctrl_enc_multiview import GeneralDITMultiviewEncoder
 
@@ -28,8 +28,3 @@ num_blocks = FADITV2Config["num_blocks"]
 FADITV2MultiCamEncoderConfig = copy.deepcopy(FADITV2Config)
 FADITV2MultiCamEncoderConfig["_target_"] = GeneralDITMultiviewEncoder
 FADITV2MultiCamEncoderConfig["layer_mask"] = [True if i > num_blocks // 2 else False for i in range(num_blocks)]
-
-num_blocks = DebugDITConfig["num_blocks"]
-DebugDITEncoderConfig = copy.deepcopy(DebugDITConfig)
-DebugDITEncoderConfig["_target_"] = GeneralDITEncoder
-DebugDITEncoderConfig["layer_mask"] = [True if i > 2 else False for i in range(num_blocks)]
