+++ cosmos_transfer1/diffusion/module/pretrained_vae.py
@@ -17,7 +17,6 @@ import os
 from abc import ABC, abstractmethod
 
 import torch
-import torch.nn.functional as F
 from einops import rearrange
 from torch.nn.modules import Module
 
@@ -604,124 +603,3 @@ class JointImageVideoSharedJITTokenizer(JointImageVideoTokenizer):
 
         self.video_vae.load_decoder(vae_dir)
         self.video_vae.load_encoder(vae_dir)
-
-
-class DummyJointImageVideoTokenizer(BaseVAE, VideoTokenizerInterface):
-    def __init__(
-        self,
-        name: str = "dummy_joint_image_video",
-        pixel_ch: int = 3,
-        latent_ch: int = 16,
-        pixel_chunk_duration: int = 17,
-        latent_chunk_duration: int = 3,
-        spatial_compression_factor: int = 16,
-        temporal_compression_factor: int = 8,
-        spatial_resolution: str = "720",
-    ):
-        self.pixel_ch = pixel_ch
-        self._pixel_chunk_duration = pixel_chunk_duration
-        self._latent_chunk_duration = latent_chunk_duration
-        self._spatial_compression_factor = spatial_compression_factor
-        self._temporal_compression_factor = temporal_compression_factor
-        self._spatial_resolution = spatial_resolution
-        super().__init__(latent_ch, name)
-
-    @property
-    def spatial_compression_factor(self):
-        return self._spatial_compression_factor
-
-    @property
-    def temporal_compression_factor(self):
-        return self._temporal_compression_factor
-
-    @property
-    def spatial_resolution(self) -> str:
-        return self._spatial_resolution
-
-    @property
-    def pixel_chunk_duration(self) -> int:
-        return self._pixel_chunk_duration
-
-    @property
-    def latent_chunk_duration(self) -> int:
-        return self._latent_chunk_duration
-
-    def get_latent_num_frames(self, num_pixel_frames: int) -> int:
-        if num_pixel_frames == 1:
-            return 1
-        assert (
-            num_pixel_frames % self.pixel_chunk_duration == 0
-        ), f"Temporal dimension {num_pixel_frames} is not divisible by chunk_length {self.pixel_chunk_duration}"
-        return num_pixel_frames // self.pixel_chunk_duration * self.latent_chunk_duration
-
-    def get_pixel_num_frames(self, num_latent_frames: int) -> int:
-        if num_latent_frames == 1:
-            return 1
-        assert (
-            num_latent_frames % self.latent_chunk_duration == 0
-        ), f"Temporal dimension {num_latent_frames} is not divisible by chunk_length {self.latent_chunk_duration}"
-        return num_latent_frames // self.latent_chunk_duration * self.pixel_chunk_duration
-
-    def encode(self, state: torch.Tensor) -> torch.Tensor:
-        B, C, T, H, W = state.shape
-        if T == 1:
-            state_B_T_C_H_W = F.interpolate(
-                rearrange(state, "b c t h w -> b t c h w"),
-                size=(self.latent_ch, H // self.spatial_compression_factor, W // self.spatial_compression_factor),
-                mode="trilinear",
-                align_corners=False,
-            )
-            return rearrange(state_B_T_C_H_W, "b t c h w -> b c t h w").contiguous()
-        assert (
-            T % self.pixel_chunk_duration == 0
-        ), f"Temporal dimension {T} is not divisible by chunk_length {self.pixel_chunk_duration}"
-        num_frames = T // self.pixel_chunk_duration * self.latent_chunk_duration
-
-        state_B_C_T_H_W = F.interpolate(
-            state,
-            size=(self.latent_ch, H // self.spatial_compression_factor, W // self.spatial_compression_factor),
-            mode="trilinear",
-            align_corners=False,
-        )
-        state_B_H_W_T_C = rearrange(state_B_C_T_H_W, "b c t h w -> b h w t c")
-        state_B_H_W_T_C = F.interpolate(
-            state_B_H_W_T_C,
-            size=(W // self.spatial_compression_factor, num_frames, self.latent_ch),
-            mode="trilinear",
-            align_corners=False,
-        )
-        return rearrange(state_B_H_W_T_C, "b h w t c -> b c t h w").contiguous()
-
-    def decode(self, latent: torch.Tensor) -> torch.Tensor:
-        B, C, T, H, W = latent.shape
-        if T == 1:
-            latent_B_T_C_H_W = F.interpolate(
-                rearrange(latent, "b c t h w -> b t c h w"),
-                size=(self.pixel_ch, H * self.spatial_compression_factor, W * self.spatial_compression_factor),
-                mode="trilinear",
-                align_corners=False,
-            )
-            return rearrange(latent_B_T_C_H_W, "b t c h w -> b c t h w").contiguous()
-
-        assert (
-            T % self.latent_chunk_duration == 0
-        ), f"Temporal dimension {T} is not divisible by chunk_length {self.latent_chunk_duration}"
-        num_frames = T * self.pixel_chunk_duration // self.latent_chunk_duration
-
-        latent_B_H_W_T_C = rearrange(latent, "b c t h w -> b h w t c")
-        latent_B_H_W_T_C = F.interpolate(
-            latent_B_H_W_T_C,
-            size=(W * self.spatial_compression_factor, num_frames, self.pixel_ch),
-            mode="trilinear",
-            align_corners=False,
-        )
-        latent_B_C_T_H_W = rearrange(latent_B_H_W_T_C, "b h w t c -> b c t h w")
-
-        state = F.interpolate(
-            latent_B_C_T_H_W,
-            size=(num_frames, H * self.spatial_compression_factor, W * self.spatial_compression_factor),
-            mode="trilinear",
-            align_corners=False,
-        )
-
-        return state.contiguous()
